Article now working through queue RabbitMQ with redis cache

+++ server/api/article/index.coffee
@@ -1,44 +1,32 @@
 amqp = require "amqplib"
 uuid = require "node-uuid"
-queueResponse = (cb) ->
-  return (req, res) ->
-    corrId = uuid()
-    responseFn = (answer, conn) ->
-      answer = JSON.parse(answer.content.toString())
-      res.json answer.status, answer.res
-      conn.close()
-    amqp.connect("amqp://localhost").then (conn) ->
-      conn.createChannel().then (ch) ->
-        ok = ch.assertQueue "",
-          exclusive: true
-        .then (qok) ->
-          qok.queue
 
-        ok = ok.then (queue) ->
-          ch.consume queue, (answer)->
-            responseFn(answer, conn)
-          ,
-            noAck: true
-          .then ->
-            queue
+conn = undefined
 
-        ok.then (queue) ->
-          queueReq = { body: req.body, params: req.params }
-          ch.sendToQueue "api.article", new Buffer( JSON.stringify( queueReq ) ),
-            correlationId: corrId
-            replyTo: queue
+amqp.connect("amqp://localhost").then (_conn) ->
+  conn = _conn
 
+if false
+  express = require("express")
+  controller = require("./article.controller")
+  router = express.Router()
+  router.get "/", controller.index
+  router.get "/:id", controller.show
+  router.post "/", controller.create
+  router.put "/:id", controller.update
+  router.patch "/:id", controller.update
+  router.delete "/:id", controller.destroy
 
-
-express = require("express")
 controller = require("./article.controller")
-router = express.Router()
-router.get "/", queueResponse controller.index
-router.get "/:id", controller.show
-router.post "/", controller.create
-router.put "/:id", controller.update
-router.patch "/:id", controller.update
-router.delete "/:id", controller.destroy
-module.exports = router
+
+queueController =
+  'api1.article.get': controller.index
+  'api1.article.get.id': controller.show
+  'api1.article.post': controller.create
+  'api1.article.put.id': controller.update
+  'api1.article.patch.id': controller.update
+  'api1.article.delete.id': controller.destroy
+
+module.exports = queueController
